Merged with va_development.
New changes this time are focused on redesigning how we use the datastore. We want to have a specification which will be east to follow and document.
We also aim to improve how objects are stored. Instead of having a key, for instance, "providers" which holds all the objects, instead we have a key for each object, for instance "providers/os_provider" and "providers/aws_provider". This works around certain race conditions and is in general a good way to store objects.

TODO - do the same with panels. Also va_development implemented a bunch of functions for adding allowed functions to users which I should make work with the new datastore.